Rename to fromFuture, make param Eval, add cats.effect.laws.util.TestInstances
