+++ laws/shared/src/test/scala/cats/effect/BaseTestsSuite.scala
@@ -16,17 +16,15 @@
 
 package cats.effect
 
-import cats.effect.laws.util.TestContext
+import cats.effect.laws.util.{TestContext, TestInstances}
 import cats.kernel.Eq
 import org.scalactic.source
 import org.scalatest.prop.Checkers
 import org.scalatest.{FunSuite, Matchers, Tag}
 import org.typelevel.discipline.Laws
 import org.typelevel.discipline.scalatest.Discipline
-import scala.concurrent.{ExecutionException, Future}
-import scala.util.{Failure, Success}
 
-class BaseTestsSuite extends FunSuite with Matchers with Checkers with Discipline {
+class BaseTestsSuite extends FunSuite with Matchers with Checkers with Discipline with TestInstances {
   /** For tests that need a usable [[TestContext]] reference. */
   def testAsync[A](name: String, tags: Tag*)(f: TestContext => Unit)
     (implicit pos: source.Position): Unit = {
@@ -47,52 +45,4 @@ class BaseTestsSuite extends FunSuite with Matchers with Checkers with Disciplin
 
   implicit def eqThrowable: Eq[Throwable] =
     Eq.fromUniversalEquals[Throwable]
-
-  /**
-   * Defines equality for `IO` references that can
-   * get interpreted by means of a [[TestContext]].
-   */
-  implicit def eqIO[A](implicit A: Eq[A], ec: TestContext): Eq[IO[A]] =
-    new Eq[IO[A]] {
-      def eqv(x: IO[A], y: IO[A]): Boolean =
-        eqFuture[A].eqv(x.unsafeToFuture(), y.unsafeToFuture())
-    }
-
-  /**
-   * Defines equality for `Future` references that can
-   * get interpreted by means of a [[TestContext]].
-   */
-  implicit def eqFuture[A](implicit A: Eq[A], ec: TestContext): Eq[Future[A]] =
-    new Eq[Future[A]] {
-      def eqv(x: Future[A], y: Future[A]): Boolean = {
-        // Executes the whole pending queue of runnables
-        ec.tick()
-
-        x.value match {
-          case None =>
-            y.value.isEmpty
-          case Some(Success(a)) =>
-            y.value match {
-              case Some(Success(b)) => A.eqv(a, b)
-              case _ => false
-            }
-          case Some(Failure(ex1)) =>
-            y.value match {
-              case Some(Failure(ex2)) =>
-                extractEx(ex1) == extractEx(ex2)
-              case _ =>
-                false
-            }
-        }
-      }
-
-      // Unwraps exceptions that got caught by Future's implementation
-      // and that got wrapped in ExecutionException (`Future(throw ex)`)
-      def extractEx(ex: Throwable): Throwable =
-        ex match {
-          case ref: ExecutionException =>
-            Option(ref.getCause).getOrElse(ref)
-          case _ => ex
-        }
-    }
 }
